Subclass Environment to simplify logic, add from_config
To make it easier to interact with the simulator, I added HuenPredictorEnvironment and RungeKuttaEnvironment classes - these implement a iterate method which iterates the physics - this is helpful as we don't need to check which implementation we're using, making this easier to extend in the future. Additionally, I added a from_config method to generate Environments from a config, which makes the usage of the class more obvious.